Module 8 assignment turn in.
Nothing works on it really. The files included are index.html, main.js
and style.css.
If I put in “trynew_data” new topojson  layer I created from a shapefile
made from just us states then the country variable becomes undefined. I
have been unable to understand why this is.
If I place in “ne_10m_admin_1_states_provinces_lakes.topojson” I get an
outline around the states, but everything on the map is blue. Also, if I
put the fill to none the graticule is behind the states, rather than on
the outside of it.
Projection specifics haven’t been fixed as I cannot get any other
projection alternatives to work. If I replace the projection with
another one it just tells me it isn’t a function.
At this point I’m pretty lost on what to do and am turning it in to get
a couple points if that.