D3 Lab turn in
It's nowhere near done, but I'm out of time to finish it. There isn't
much more done than what was done module 9 beyond some changes to the
code as I was making my way through the module. No interactivity beyond
the drop down bar and map changes when one of the options is clicked.
There isn't much here to grade, but this contains the mainlab9 js file
and the style.css file.