D3 turn in
Waiting to hand it in to work on it didn't allow me to get much further.
There isn't really much here to grade. It contains the index.html,
main.js and style.css. The map doesn't really do anything, it changes
when drop down options are changed. It reacts to the stands being panned
over, however it activates the wrong parts of the bar chart. There are
no finished titles. Highlight turns on everything is scrolls over but
does not turn off. If there is really anything done here correctly or
well I'm not sure what it is, but I needed to turn something in. I
figured any amount of points would be better than getting absolutely
zero.