d3 hand in main change
I made a slight change to some code pieces in the main so I'm
reuploading it. My comments from the last commit still stand, but I just
wanted it to be fixed when I saw there was an issue I could confront.
The index and style sheet for it are the same and where included in the
comit to keep them together.